requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
pdfkit==1.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pypika-tortoise==0.1.6
pytz==2023.3.post1
six==1.16.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
tzlocal==5.1
yarl==1.9.2
